refactor: moving I/O back into operations.rs
